refactor(ui): reorganize project components into desktop module
- Move project-related UI components from `desktop-shared` to `desktop` package to better align with module boundaries.
- Update `NavigationSidebar` and `SessionActionsMenu` to reflect moved dependencies and remove obsolete project dialogs from shared scope.
- Add `spaceId` and `spaceName` fields to `Project` model and repository for team edition support.
- Alter `projects` database table to include new `space_id` and `space_name` columns.
- Adjust `SkillsHistoryPanel` timestamp formatting to use the system default timezone instead of raw UTC.
- Implement a fallback in `InstantSerializer` to handle legacy timestamps lacking the 'Z' suffix.
