+++ desktop/src/main/kotlin/io/askimo/desktop/project/KnowledgeSourceModels.kt
@@ -2,7 +2,7 @@
  *
  * Copyright (c) 2026 Askimo
  */
-package io.askimo.ui.project
+package io.askimo.desktop.project
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.InsertDriveFile
