+++ desktop/src/main/kotlin/io/askimo/desktop/project/ProjectViewModel.kt
@@ -2,7 +2,7 @@
  *
  * Copyright (c) 2026 Askimo
  */
-package io.askimo.ui.project
+package io.askimo.desktop.project
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
