+++ shared/src/main/kotlin/io/askimo/core/db/DatabaseManager.kt
@@ -201,6 +201,22 @@ class DatabaseManager private constructor(
             } catch (_: Exception) {
                 // Column already exists — safe to ignore.
             }
+
+            try {
+                stmt.executeUpdate(
+                    "ALTER TABLE projects ADD COLUMN space_id TEXT",
+                )
+            } catch (_: Exception) {
+                // Column already exists — safe to ignore.
+            }
+
+            try {
+                stmt.executeUpdate(
+                    "ALTER TABLE projects ADD COLUMN space_name TEXT",
+                )
+            } catch (_: Exception) {
+                // Column already exists — safe to ignore.
+            }
         }
     }
